requirements.txt
BetBot
requests>=2.18.4
python-dateutil>=2.6.1
pytz>=pytz-2017.2